requirements.txt
fastapi
python-multipart
uvicorn[standard]==0.15.0
gunicorn==20.1.0
onnxruntime-gpu
torch
transformers
tqdm
elasticsearch==7.17.
sentence_transformers